requirements.txt

asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
filelock==3.16.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
numpy==2.1.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
pipenv==2024.4.0
platformdirs==4.3.6
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
selenium==4.26.1
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
websocket-client==1.8.0
whitenoise==6.8.2
wsproto==1.2.0
